renamed models main file to index.js
